requirements.txt
aenum==3.1.15
aiohttp==3.9.1
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.3
Deprecated==1.2.14
distro==1.9.0
Flask==3.0.1
flask-restx==1.3.0
frozenlist==1.4.1
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
line-bot-sdk==3.7.0
MarkupSafe==2.1.4
multidict==6.0.5
openai==1.10.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
